+++ libs/core/src/components/dropdown/dropdown.ts
@@ -2,7 +2,9 @@ import { property, query, queryAsync, state } from 'lit/decorators.js'
 import { unsafeHTML } from 'lit/directives/unsafe-html.js'
 import { when } from 'lit/directives/when.js'
 import { ifDefined } from 'lit/directives/if-defined.js'
+import { classMap } from 'lit/directives/class-map.js'
 import { msg, str, updateWhenLocaleChanges } from '@lit/localize'
+import { HTMLTemplateResult } from 'lit'
 import 'reflect-metadata'
 
 import { constrainSlots } from '../../utils/helpers'
@@ -26,7 +28,6 @@ import { GdsFormControlElement } from '../form-control'
 
 import styles from './dropdown.styles'
 import { TransitionalStyles } from '../../utils/helpers/transitional-styles'
-import { CSSResult, HTMLTemplateResult } from 'lit'
 
 /**
  * @element gds-dropdown
@@ -120,8 +121,21 @@ export class GdsDropdown<ValueT = any>
    * the popover width. If you use this option, make sure to verify that your options
    * are still readable and apply appropriate custom layout or truncation if neccecary.
    */
-  @property({ type: Boolean })
+  @property({ type: Boolean, attribute: 'sync-popover-width' })
   syncPopoverWidth = false
+
+  /**
+   * Size of the dropdown. Supports `medium` and `small`. There is no `large` size for dropdowns.
+   * `medium` is the default size.
+   */
+  @property()
+  size: 'medium' | 'small' = 'medium'
+
+  /**
+   * Whether to hide the label.
+   */
+  @property({ type: Boolean, attribute: 'hide-label' })
+  hideLabel = false
 
   /**
    * Get the options of the dropdown.
@@ -210,7 +224,7 @@ export class GdsDropdown<ValueT = any>
     return html`
       ${this._tStyles}
       ${when(
-        this.label,
+        this.label && !this.hideLabel,
         () => html`<label for="trigger">${this.label}</label>`
       )}
 
@@ -224,6 +238,8 @@ export class GdsDropdown<ValueT = any>
         aria-owns="listbox"
         aria-controls="listbox"
         aria-expanded="${this.open}"
+        aria-label="${this.label}"
+        class=${classMap({ small: this.size === 'small' })}
       >
         <slot name="trigger">
           <span>${unsafeHTML(this.displayValue)}</span>